feat(views): :sparkles: update my-job-applications.blade.php to handle empty job applications
The code changes in `my-job-applications.blade.php` add a conditional statement to handle the case when there are no job applications found. It displays a message indicating that no job applications were found. This improves the user experience by providing clear feedback when there are no job applications to display.