refactor(views): :recycle: update my-jobs.blade.php to display the number of job applications
This commit modifies the my-jobs.blade.php file to display the number of job applications for each job. The change is made by replacing the static "0 Applications" text with the dynamic count of applications retrieved from the $job->applications relationship. This improvement provides users with valuable information about the number of applications received for each job.