refactor(views): :recycle: update job create and edit forms
This commit refactors the job create and edit forms in `create.blade.php` and `edit.blade.php` respectively. It replaces the plain textareas with a new class `text-editor` to enable the Trumbowyg WYSIWYG editor for the description, benefits, responsibility, and qualifications fields. This enhancement improves the user experience by providing a more intuitive and user-friendly way to input and format text in these fields.