feat(middleware): :sparkles: add Admin middleware
This commit adds the Admin middleware, which handles incoming requests and checks if the authenticated user has the 'admin' role. If the user does not have the 'admin' role, they are redirected to the profile page with an error message. This addition improves the security and access control of the application.