feat(controllers): :sparkles: add UserController with index method
This commit adds the UserController to the Admin namespace. The controller includes an index method that retrieves users from the User model and paginates them. The retrieved users are then passed to the 'admin.users.show-list' view. This addition improves the organization and structure of the codebase by separating the user-related logic into its own controller.