feat(views): :sparkle: add delete functionality to user dropdown menu
This commit adds the functionality to delete a user from the user dropdown menu in the admin users list view. When the delete button is clicked, a confirmation dialog is shown, and if confirmed, an AJAX request is sent to delete the user. Upon successful deletion, the page is redirected to the admin users list view, improving the user management experience.