feat(controllers): :sparkles: implement store method for new user creation from admin view
This commit implements the store method that is responsible for creating a new user from admin view. The medhod first validates the form data from the http request came from AJAX. If the form data validation passes (or not), then a new user is created and a session flash message is shown.