feat(sparkles): :sparkles: add reset password functionality to AccountController
This commit adds the resetPassword method to the AccountController, allowing users to reset their passwords using a password reset token. The method retrieves the token from the database and checks its validity. If the token is valid, the user is redirected to the reset password page with the token as a parameter. If the token is invalid, an error message is displayed. This change improves the security and usability of the application.